refactor and tidy up
